requirements.txt
python-dotenv==0.21.1
langchain==0.0.189
openai==0.27.4
unstructured
tabulate
chromadb
tqdm
transformers